[Online book Store Project]

+++ onlineBookStore/src/main/java/com/bookstore/onlineBookStore/repository/BookStoreRepository.java
@@ -4,5 +4,5 @@ import com.bookstore.onlineBookStore.entity.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @org.springframework.stereotype.Repository
-public interface BookStoreRepository extends JpaRepository<Book, Long> {
+public interface BookStoreRepository extends JpaRepository<Book, Integer> {
 }
